Pill renaming bloc related classes
