+++ pill/lib/bloc/pill_state.dart
