+++ pill/lib/widget/pillWidget.dart
@@ -1,7 +1,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:pill/bloc/PillEvent.dart';
+import 'package:pill/bloc/pilll_event.dart';
 import 'package:pill/bloc/pill_bloc.dart';
 import 'package:pill/model/PillToTake.dart';
